+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/bullhorn/novo-elements/compare/v0.0.5...v0.0.6) (2016-05-05)
+
+
+### Bug Fixes
+
+* **colors:** Hooks up toast to color copy func([6bbd9a6](https://github.com/bullhorn/novo-elements/commit/6bbd9a6))
+* **style:** fixes href cursor styling ([#39](https://github.com/bullhorn/novo-elements/issues/39))([82ca7b6](https://github.com/bullhorn/novo-elements/commit/82ca7b6)), closes [#39](https://github.com/bullhorn/novo-elements/issues/39)
+
+
+### Features
+
+* **fork me:** Adds 'Fork Me On GH' button to side nav, cleans side nav styles ([#37](https://github.com/bullhorn/novo-elements/issues/37))([ecf4010](https://github.com/bullhorn/novo-elements/commit/ecf4010))
+* **list:** Adds list component and documentation, closes [#11](https://github.com/bullhorn/novo-elements/issues/11) ([#40](https://github.com/bullhorn/novo-elements/issues/40))([3e11da6](https://github.com/bullhorn/novo-elements/commit/3e11da6)), closes [#11](https://github.com/bullhorn/novo-elements/issues/11) [#40](https://github.com/bullhorn/novo-elements/issues/40)
+
+
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/bullhorn/novo-elements/compare/v0.0.4...v0.0.5) (2016-05-03)
+
+
+### Features
+
+* **dropdown:** Adding Dropdown Component ([#33](https://github.com/bullhorn/novo-elements/issues/33)), closes [#9](https://github.com/bullhorn/novo-elements/issues/9)([411ab43](https://github.com/bullhorn/novo-elements/commit/411ab43)), closes [#9](https://github.com/bullhorn/novo-elements/issues/9)
+* **header:** Adds header element & documentation, closes [#10](https://github.com/bullhorn/novo-elements/issues/10)([c565d0e](https://github.com/bullhorn/novo-elements/commit/c565d0e)), closes [#10](https://github.com/bullhorn/novo-elements/issues/10)
+
+
+
+<a name="0.0.4"></a>
+## [0.0.4](https://github.com/bullhorn/novo-elements/compare/v0.0.3...v0.0.4) (2016-05-03)
+
+
+### Features
+
+* **drawer:** Add drawer element, closes [#4](https://github.com/bullhorn/novo-elements/issues/4)([2a16f07](https://github.com/bullhorn/novo-elements/commit/2a16f07)), closes [#4](https://github.com/bullhorn/novo-elements/issues/4)
+* **select:** Added Select element, documentation, and initial tests([f1036d2](https://github.com/bullhorn/novo-elements/commit/f1036d2))
+* **switch:** Adds switch element and documentation, closes [#16](https://github.com/bullhorn/novo-elements/issues/16)([747174b](https://github.com/bullhorn/novo-elements/commit/747174b)), closes [#16](https://github.com/bullhorn/novo-elements/issues/16)
+* **toast:** Adds toast element and documentation, closes [#18](https://github.com/bullhorn/novo-elements/issues/18)([08f2351](https://github.com/bullhorn/novo-elements/commit/08f2351)), closes [#18](https://github.com/bullhorn/novo-elements/issues/18)
+* **tooltip:** Adding tooltip feature, closes [#20](https://github.com/bullhorn/novo-elements/issues/20) ([#29](https://github.com/bullhorn/novo-elements/issues/29))([211f730](https://github.com/bullhorn/novo-elements/commit/211f730)), closes [#20](https://github.com/bullhorn/novo-elements/issues/20) [(#29](https://github.com/(/issues/29)
+
+
+
 <a name="0.0.3"></a>
 ## [0.0.3](https://github.com/bullhorn/novo-elements/compare/v0.0.2...v0.0.3) (2016-04-28)